% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkSet.R
\name{BenchmarkSet}
\alias{BenchmarkSet}
\title{Benchmark Set}
\description{
Benchmark Set

Benchmark Set
}
\details{
R interface to a YAHPO GYM BenchmarkSet.


Allows exporting the objective function from a `YAHPO GYM` BenchmarkSet.
and additional helper functionality.
}
\section{Methods}{

  * new(key, onnx_session): Initialize the class.
  * get_objective_function(): Obtain the [`bbotk::ObjectiveFunction`].
  * get_param_set(): Obtain the [`paradox::ParamSet`].
}

\section{Fields}{

  * session: [`onnx.InferenceSession`] \cr
  * instances: [`character`] \cr
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BenchmarkSet$new()}}
\item \href{#method-get_objective_function}{\code{BenchmarkSet$get_objective_function()}}
\item \href{#method-eval_objective_function}{\code{BenchmarkSet$eval_objective_function()}}
\item \href{#method-get_opt_space}{\code{BenchmarkSet$get_opt_space()}}
\item \href{#method-clone}{\code{BenchmarkSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$new(key, onnx_session = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objective_function"></a>}}
\if{latex}{\out{\hypertarget{method-get_objective_function}{}}}
\subsection{Method \code{get_objective_function()}}{
Get the objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_objective_function(instance, drop_fidelity_params)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A [`bbotk::Objective`].
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval_objective_function"></a>}}
\if{latex}{\out{\hypertarget{method-eval_objective_function}{}}}
\subsection{Method \code{eval_objective_function()}}{
Evaluate the objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$eval_objective_function(xs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{xs}}{[`instance`] \cr
A valid configuration. See `get_opt_param_set`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector, prediction results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_opt_space"></a>}}
\if{latex}{\out{\hypertarget{method-get_opt_space}{}}}
\subsection{Method \code{get_opt_space()}}{
Get Optimization Param Set
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_opt_space(instance, drop_fidelity_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{[`instance`] \cr
A valid instance. See `instances`.}

\item{\code{drop_fidelity_params}}{[`logical`] \cr
Should fidelity params be dropped?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`paradox::ParamSet`] containing the search space to optimize over.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
