% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tasks.R
\name{get_tasks}
\alias{get_tasks}
\title{Get versioned task sets}
\usage{
get_tasks(type, version = 0L)
}
\arguments{
\item{type}{[`character`] \cr
Type of task set to get. For now 'single' (single-crit) and 'multi' (multi-crit) are supported.}

\item{version}{[`integer`] \cr
Version of the task set to get. Default is 0.}
}
\value{
A data.frame containing a list of tasks.
}
\description{
Get Optimization ConfigSpace
}
