% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_local_config}
\alias{init_local_config}
\title{Initialize a local configuration}
\usage{
init_local_config(data_path = NULL, settings_path = "~/.config/yahpo_gym")
}
\arguments{
\item{data_path}{`character`\cr
Path to save meta-data to.}

\item{settings_path}{`character`\cr
Path to save configuration object including "path" to.}
}
\description{
Initialize a local configuration
}
\details{
Allows setting a path the required meta-data for YAHPO Gym is installed to.
}
