% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{preproc_xs}
\alias{preproc_xs}
\title{Preprocess r object for use with python's YAHPO Gym}
\usage{
preproc_xs(xss, ...)
}
\arguments{
\item{...}{`any` \cr
Named params, appended to `xs`.}

\item{xs}{`list` of `list` \cr
List of hyperparams}
}
\description{
Preprocess r object for use with python's YAHPO Gym
}
