% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\docType{package}
\name{yahpogym-package}
\alias{yahpogym}
\alias{yahpogym-package}
\title{yahpogym: Interface to the yahpo gym python module}
\description{
A package that connects YAHPO GYM to R.
}
\author{
\strong{Maintainer}: Florian Pfisterer \email{florian.pfisterer@stat.uni-muenchen.de}

}
