% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkSet.R
\name{BenchmarkSet}
\alias{BenchmarkSet}
\title{Benchmark Set}
\description{
Benchmark Set

Benchmark Set
}
\details{
R interface to a YAHPO GYM BenchmarkSet.


Allows exporting the objective function from a `YAHPO GYM` BenchmarkSet.
and additional helper functionality.
}
\section{Methods}{

  * new(key, onnx_session): Initialize the class.
  * get_objective_function(): Obtain the [`bbotk::ObjectiveFunction`].
  * get_param_set(): Obtain the [`paradox::ParamSet`].
}

\section{Fields}{

  * session: [`onnx.InferenceSession`] \cr
  * instances: [`character`] \cr
}

\examples{
\dontrun{
b = BenchmarkSet$new("lcbench")
b$instances
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_instance}}{[`BenchmarkSet`] \cr
A python `yahpo_gym.BenchmarkSet`.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{session}}{`onnxruntime.InferenceSession` \cr
Set/Get the ONNX session.

The param `sess` is a `onnxruntime.InferenceSession`\cr
  A matching `onnxruntime.InferenceSession`. Please make sure
  that the session matches the selected benchmark scenario, as no
  additional checks are performed and inference will fail during
  evaluation of the objective function.}

\item{\code{instances}}{`character` \cr
A character vector of available instances for the scenario.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BenchmarkSet$new()}}
\item \href{#method-get_objective}{\code{BenchmarkSet$get_objective()}}
\item \href{#method-get_opt_space_py}{\code{BenchmarkSet$get_opt_space_py()}}
\item \href{#method-clone}{\code{BenchmarkSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$new(
  key,
  onnx_session = NULL,
  download = TRUE,
  active_session = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{`character` \cr
Key for a benchmark scenario. See [`list_benchmarks`] for more information.}

\item{\code{onnx_session}}{`onnxruntime.InferenceSession` \cr
A matching `onnxruntime.InferenceSession`. See `session` for more information.
If no session is provided, new session is created.}

\item{\code{download}}{`logical` \cr
Download the required data on instantiation? Default `TRUE`.}

\item{\code{active_session}}{`logical` \cr
Should the benchmark run in an active `onnxruntime.InferenceSession`? Initialized to `FALSE`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objective"></a>}}
\if{latex}{\out{\hypertarget{method-get_objective}{}}}
\subsection{Method \code{get_objective()}}{
Get the objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_objective(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{[`instance`] \cr
A valid instance. See `instances`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`Objective`][bbotk::Objective] containing "domain", "codomain" and a
 functionality to evaluate the surrogates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_opt_space_py"></a>}}
\if{latex}{\out{\hypertarget{method-get_opt_space_py}{}}}
\subsection{Method \code{get_opt_space_py()}}{
Get Optimization ConfigSpace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_opt_space_py(instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{[`instance`] \cr
A valid instance. See `instances`.}

\item{\code{drop_fidelity_params}}{[`logical`] \cr
Should fidelity params be dropped?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`paradox::ParamSet`] containing the search space to optimize over.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
