% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkSet.R
\name{BenchmarkSet}
\alias{BenchmarkSet}
\title{Benchmark Set}
\description{
Benchmark Set

Benchmark Set
}
\details{
R interface to a YAHPO GYM BenchmarkSet.


Allows exporting the objective function from a `YAHPO GYM` BenchmarkSet.
and additional helper functionality.
}
\section{Methods}{

  * new(key, onnx_session, active_session, download): Initialize the class.
  * get_objective(): Obtain the [`bbotk::Objective`].
  * get_opt_space_py(): Obtain the [`ConfigSpace`].
}

\section{Fields}{

  * session: [`onnx.InferenceSession`] \cr
  * instances: [`character`] \cr
}

\examples{
\dontrun{
b = BenchmarkSet$new("lcbench")
b$instances
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{`character` \cr
The scenario identifier.}

\item{\code{onnx_session}}{`onnxruntime.InferenceSession` \cr
A session to use for the predict. If `NULL` a new session is initialized.}

\item{\code{download}}{`logical` \cr
Download data in case it is not available?}

\item{\code{check}}{`logical` \cr
Check whether values coincide with `domain`.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{session}}{`onnxruntime.InferenceSession` \cr
Set/Get the ONNX session.

The param `sess` is a `onnxruntime.InferenceSession`\cr
  A matching `onnxruntime.InferenceSession`. Please make sure
  that the session matches the selected benchmark scenario, as no
  additional checks are performed and inference will fail during
  evaluation of the objective function.}

\item{\code{instances}}{`character` \cr
A character vector of available instances for the scenario.}

\item{\code{domain}}{`ParamSet` \cr
A [`paradox::ParamSet`] describing the domain to be optimized over.}

\item{\code{codomain}}{`ParamSet` \cr
A [`paradox::ParamSet`] describing the output domain.}

\item{\code{quant}}{`numeric` \cr
Multiply runtime by this factor. Defaults to 0.01.}

\item{\code{py_instance}}{[`BenchmarkSet`] \cr
A python `yahpo_gym.BenchmarkSet`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BenchmarkSet$new()}}
\item \href{#method-get_objective}{\code{BenchmarkSet$get_objective()}}
\item \href{#method-get_opt_space_py}{\code{BenchmarkSet$get_opt_space_py()}}
\item \href{#method-subset_codomain}{\code{BenchmarkSet$subset_codomain()}}
\item \href{#method-clone}{\code{BenchmarkSet$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$new(
  key,
  onnx_session = NULL,
  active_session = FALSE,
  download = FALSE,
  check = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{`character` \cr
Key for a benchmark scenario. See [`list_benchmarks`] for more information.}

\item{\code{onnx_session}}{`onnxruntime.InferenceSession` \cr
A matching `onnxruntime.InferenceSession`. See `session` for more information.
If no session is provided, new session is created.}

\item{\code{active_session}}{`logical` \cr
Should the benchmark run in an active `onnxruntime.InferenceSession`? Initialized to `FALSE`.}

\item{\code{download}}{`logical` \cr
Download the required data on instantiation? Default `FALSE`.}

\item{\code{check}}{`logical` \cr
Check inputs for validity before passing to surrogate model? Default `FALSE`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_objective"></a>}}
\if{latex}{\out{\hypertarget{method-get_objective}{}}}
\subsection{Method \code{get_objective()}}{
Get the objective function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_objective(
  instance,
  check_values = FALSE,
  timed = FALSE,
  logging = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{[`instance`] \cr
A valid instance. See `instances`.}

\item{\code{check_values}}{(`logical`) \cr
Should values be checked by bbotk? Initialized to `FALSE`.}

\item{\code{timed}}{(`logical`) \cr
Should function evaluation simulate runtime? Initialized to `FALSE`.}

\item{\code{logging}}{(`logical`) \cr
Should function evaluationd be logged? Initialized to `FALSE`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`Objective`][bbotk::Objective] containing "domain", "codomain" and a
 functionality to evaluate the surrogates.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_opt_space_py"></a>}}
\if{latex}{\out{\hypertarget{method-get_opt_space_py}{}}}
\subsection{Method \code{get_opt_space_py()}}{
Get Optimization ConfigSpace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$get_opt_space_py(instance, drop_fidelity_params = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{instance}}{[`instance`] \cr
A valid instance. See `instances`.}

\item{\code{drop_fidelity_params}}{[`logical`] \cr
Should fidelity params be dropped? Defaults to `FALSE`.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`paradox::ParamSet`] containing the search space to optimize over.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset_codomain"></a>}}
\if{latex}{\out{\hypertarget{method-subset_codomain}{}}}
\subsection{Method \code{subset_codomain()}}{
Subset the codomain
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$subset_codomain(keep)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keep}}{(`character`) \cr
Vector of co-domain target names to keep.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A [`paradox::ParamSet`] containing the output space (codomain).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkSet$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
