% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objective.R
\name{preproc_xs}
\alias{preproc_xs}
\title{Preprocess r object for use with python's YAHPO GYM}
\usage{
preproc_xs(xs, ...)
}
\arguments{
\item{xs}{`list` \cr
List of hyperparams}

\item{...}{`any` \cr
Named params, appended to `xs`.}
}
\description{
Preprocess r object for use with python's YAHPO GYM
}
