% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{yahpogym-package}
\alias{yahpogym}
\alias{yahpogym-package}
\title{yahpogym: Interface to the YAHPO Gym python module}
\description{
A package that connects YAHPO Gym to R.
}
\author{
\strong{Maintainer}: Florian Pfisterer \email{florian.pfisterer@stat.uni-muenchen.de}

Authors:
\itemize{
  \item Lennart Schneider \email{lennart.sch@web.de}
}

}
